+++ src/api/mod.rs
@@ -1,0 +1,4 @@
+// rest server
+pub mod rest;
+
+pub use rest::host_rest_api_server;