+++ src/error.rs
@@ -1,21 +1,26 @@
+use crate::api::rest::ApiError;
 pub use eyre::Error as EyreError;
-pub use thiserror::Error;
+pub use std::io::Error as IoError;
+pub use thiserror::Error as ThisError;
+pub use toml::de::Error as TomlError;
 
-#[derive(Debug, Error)]
+#[derive(Debug, ThisError)]
 pub enum ConfigError {
     #[error("Failed to read config file: {0}")]
     Io(#[from] std::io::Error),
     #[error("Failed to parse TOML: {0}")]
-    Toml(#[from] toml::de::Error),
+    Toml(#[from] TomlError),
 }
 
 // TODO implement chaining errors
-#[derive(Debug, Error)]
+#[derive(Debug, ThisError)]
 pub enum SignalsError {
     #[error("generic error: {0}")]
     Generic(EyreError),
     #[error("conifg error: {0}")]
     Config(ConfigError),
+    #[error("api error: {0}")]
+    Api(ApiError),
 }
 
 pub type Result<T> = std::result::Result<T, SignalsError>;
@@ -29,5 +34,11 @@ impl From<EyreError> for SignalsError {
 impl From<ConfigError> for SignalsError {
     fn from(error: ConfigError) -> Self {
         SignalsError::Config(error)
+    }
+}
+
+impl From<ApiError> for SignalsError {
+    fn from(error: ApiError) -> Self {
+        SignalsError::Api(error)
     }
 }