+++ src/lib.rs
@@ -1,4 +1,10 @@
-// custom error
-pub mod error;
-// configuration
+// third party api
+pub mod adapter;
+// api server
+pub mod api;
+// toml configuration
 pub mod config;
+// project custom error
+pub mod error;
+// metadata
+pub mod metadata;