+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # [signal](./README.md) changelog
 
+## [0.1.1]
+### Added
+- REST endpoint
+- custom error
+- TOML config
+- runner management
+- coinmarketcap runner
+
 ## [0.1.0]
 ### Added
 - hello world